Fix indexed zero page addressing
When an instruction uses the indexed zero page addressing mode, if the
sum of the index and the address is larger than the zero page, the
address will wrap around inside the zero page. For example, and address
0xFA and an index of 0x06 would result in 0x101, but the actual address
with be 0x01.

+++ 6502_utils.h
@@ -77,7 +77,9 @@ Word getZeroPageAddress() {
 }
 
 Word getZeroIndexedAddress(Byte index) {
-    return (0 << 8) | (fetchInstruction() + index);
+    // Must truncate lower Byte because indexed
+    // zero page address wraps inside zero page
+    return (0 << 8) | (Byte)(fetchInstruction() + index);
 }
 
 Word getIndexedIndirectAddress(Byte index) {
